fix: prevent PrismaClient from being bundled into browser
Two import chains were pulling server-only PrismaClient into client bundles:

1. providers.tsx → use-branding.ts → system-settings.ts → db.ts
   Fix: Extract DEFAULT_BRANDING into src/lib/branding.ts (no db import)

2. sidebar-content.tsx → permissions/index.ts → permissions/check.ts → db.ts
   Fix: Remove check.ts and create-default-roles.ts re-exports from barrel;
   server files import directly from @/lib/permissions/check

Also fixes db.ts $extends type cast (use `as any` for Prisma's complex generics).
