+++ src/app/api/projects/[projectId]/roles/[roleId]/route.ts
@@ -3,7 +3,8 @@ import { z } from 'zod'
 import { handleApiError, notFoundError } from '@/lib/api-utils'
 import { requireAuth, requirePermission, requireProjectByKey } from '@/lib/auth-helpers'
 import { db } from '@/lib/db'
-import { isMember, isValidPermission, PERMISSIONS, parsePermissions } from '@/lib/permissions'
+import { isValidPermission, PERMISSIONS, parsePermissions } from '@/lib/permissions'
+import { isMember } from '@/lib/permissions/check'
 
 // Schema for updating a role
 const updateRoleSchema = z.object({
